requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
dataclasses-json==0.6.3
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.104.1
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.0.354
langchain-community==0.0.11
langchain-core==0.1.8
langchain-google-genai==0.0.5
langsmith==0.0.78
marshmallow==3.20.1
matplotlib-inline==0.1.6

multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.3
openai==1.7.1
packaging==23.2
pandas==2.1.4
parso==0.8.3
pickleshare==0.7.5
pip==23.3.1
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.42
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.0
pure-eval==0.2.2
py-youtube==1.1.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.20.0
pydantic==1.10.13
Pygments==2.17.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
pywin32
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.3
tomli==2.0.1
tornado==6.2
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
wcwidth==0.2.13
websockets==12.0
wheel==0.41.2
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.17.0
